reworked the widgets
removed volumenbar since it no longer works and I never noticed.

added a battery widget to show the state of the battery

splitted the freq and temperature view since my laptop has no
tempearature value

systl widgets now test first if the sysctl value is available. If not
the widget will not be visible and added a view lines to test if a
widget s visble from the start to remove if so.