Dynamic loading of sysctl widgets and update to temp
Widget which use sysctl now first check if they can access there sysctl
values. If not the rc.lua will just not add it into the taskbar.

temp now has a popup which displays the averages over the last 10, 100 and 1000 values,
as well as the min and max overall and over all values in the history
buffer.